switch from inline declaration-plus-initialization of class static ...
... variables to employing a static class init method that's defined
in the class and then called at the end of the file, outside the class.

This is so that some Gooble code, which can't handle something or other
for now but will be able to later, doesn't become incontinent on Aunt
Gertrude's davenport.

But, honestly, it's probably better discipline to have all the
class-static vars collected in one place; though too having to call
the init function, and always remember to do so, at the bottom of
the file ain't pretty.