establish cursor provenance-based go/no-go (guest/exile) list: ...
... use e.g. AppendToCursorPrefixGuestlist("swine") to make tracked
piggies show up as cursors. Note that when the guest list is empty
all are allowed; the guest list is run first, then the exile list,
with the result that if there's a guest list in the first place
the exile list can only narrow it; whereas if there's no guest list
the exile list subtracts from otherwise infinite admissibility.