update sample code to assume server-based operation (because now it's ...
... more or less unworkable to run local code directly in a browser,
via timesomely predictable bastardry), by updating 'path' references.

Note that for this to run at this exact moment in the commit history,
you have to symlink zeugma into the node_modules directory, as the
package.json specification doesn't pull zeugma in from any distant
place. This state of being is, to be clear, desirable: dynamic
development of the library and the sample code that exercises it can
proceed back and forth without any network (npm) nonsense.